+++ ndll/pipeline/operators/hybrid_jpg_decoder_test.cc
@@ -135,7 +135,7 @@ TYPED_TEST(HybridDecoderTest, JPEGDecode) {
   Pipeline pipe(
       batch_size,
       num_thread,
-      main_stream,
+      (int64)main_stream,
       0);
 
   shared_ptr<Batch<CPUBackend>> batch(CreateJPEGBatch<CPUBackend>(
